+++ src/helpers/Component.php
@@ -77,6 +77,27 @@ class Component
     }
 
     /**
+     * Cleanses a component config of any `on X` or `as X` keys.
+     *
+     * @param array $config
+     * @return array
+     * @since 3.9.4
+     */
+    public static function cleanseConfig(array $config): array
+    {
+        foreach ($config as $key => $value) {
+            if (is_string($key) && (strncmp($key, 'on ', 3) === 0 || strncmp($key, 'as ', 3) === 0)) {
+                unset($config[$key]);
+                continue;
+            }
+            if (is_array($value)) {
+                $config[$key] = static::cleanseConfig($value);
+            }
+        }
+        return $config;
+    }
+
+    /**
      * Instantiates and populates a component, and ensures that it is an instance of a given interface.
      *
      * @param mixed $config The component’s class name, or its config, with a `type` value and optionally a `settings` value.
@@ -108,7 +129,7 @@ class Component
 
         // Instantiate and return
         $config['class'] = $class;
-        return Craft::createObject($config);
+        return Craft::createObject(static::cleanseConfig($config));
     }
 
     /**